css/css-overscroll-behavior/parsing/overscroll-behavior-shorthand.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Overscroll Behavior: overscroll-behavior sets longhands</title>
<link rel="help" href="https://drafts.csswg.org/css-overscroll-behavior/#propdef-overscroll-behavior">
<meta name="assert" content="overscroll-behavior supports the full grammar '[ contain | none | auto ]{1,2}'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/shorthand-testcommon.js"></script>
</head>
<body>
<script>
test_shorthand_value('overscroll-behavior', 'contain', {
  'overscroll-behavior-x': 'contain',
  'overscroll-behavior-y': 'contain'
});

test_shorthand_value('overscroll-behavior', 'none auto', {
  'overscroll-behavior-x': 'none',
  'overscroll-behavior-y': 'auto'
});
</script>
</body>
</html>
